requirements.txt
langchain>=0.3
langchain-google-genai>=2.0
langgraph>=0.2
python-dotenv>=1.0
pandas>=2.0
streamlit>=1.30
